siphon off test data
