+++ channel/state/outcome/allocation_test.go
@@ -8,38 +8,6 @@ import (
 	"github.com/statechannels/go-nitro/types"
 )
 
-var testAllocations = Allocations{{
-	Destination:    types.Destination(common.HexToHash("0x00000000000000000000000096f7123E3A80C9813eF50213ADEd0e4511CB820f")),
-	Amount:         big.NewInt(1),
-	AllocationType: 0,
-	Metadata:       zeroBytes}}
-
-var allocsX = Allocations{ // [{Alice: 2, Bob: 3}]
-	{
-		Destination:    alice,
-		Amount:         big.NewInt(2),
-		AllocationType: 0,
-		Metadata:       make(types.Bytes, 0)},
-	{
-		Destination:    bob,
-		Amount:         big.NewInt(3),
-		AllocationType: 0,
-		Metadata:       make(types.Bytes, 0)},
-}
-
-var allocsY = Allocations{ // [{Bob: 2, Alice: 1}]
-	{
-		Destination:    bob,
-		Amount:         big.NewInt(2),
-		AllocationType: 0,
-		Metadata:       make(types.Bytes, 0)},
-	{
-		Destination:    alice,
-		Amount:         big.NewInt(1),
-		AllocationType: 0,
-		Metadata:       make(types.Bytes, 0)},
-}
-
 func TestEqualAllocations(t *testing.T) {
 
 	var a1 = Allocations{{ // [{Alice: 2}]
