+++ channel/state/outcome/exit_test.go
@@ -2,7 +2,6 @@ package outcome
 
 import (
 	"bytes"
-	"encoding/hex"
 	"fmt"
 	"math/big"
 	"testing"
@@ -106,13 +105,6 @@ func TestEqualExits(t *testing.T) {
 	}
 }
 
-var zeroBytes = []byte{}
-
-var testExit = Exit{{Asset: common.HexToAddress("0x00"), Metadata: zeroBytes, Allocations: testAllocations}}
-
-// copy-pasted from https://github.com/statechannels/exit-format/blob/201d4eb7554bac337a780cc8a640f6c45c3045a5/test/exit-format-ts.test.ts
-var encodedExitReference, _ = hex.DecodeString("00000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000060000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000002000000000000000000000000096f7123e3a80c9813ef50213aded0e4511cb820f0000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000000")
-
 func TestExitEncode(t *testing.T) {
 	var encodedExit, err = testExit.Encode()
 
@@ -134,22 +126,6 @@ func TestExitDecode(t *testing.T) {
 	if !testExit.Equal(decodedExit) {
 		t.Error("decoded exit does not match expectation")
 	}
-}
-
-var alice = types.Destination(common.HexToHash("0x0a"))
-var bob = types.Destination(common.HexToHash("0x0b"))
-
-var e Exit = Exit{
-	{
-		Asset:       types.Address{}, // eth, fil, etc.
-		Metadata:    zeroBytes,
-		Allocations: allocsX,
-	},
-	{
-		Asset:       types.Address{123}, // some token
-		Metadata:    zeroBytes,
-		Allocations: allocsY,
-	},
 }
 
 func TestTotal(t *testing.T) {
